Updates lo let user mint NFTs
every user needs an own collection with its own candy machine, so I implemented a logic which allow every user that never created a project to create its own collection first and map the user wallet with the address on a solana program written with Anchor (see hashmap_solana_program.js and idl.json which are used to communicate with the deployed solana program).
More updates for testing purposes.